Refactoring to align better with modern software standards
- Refactoring to use SQLite instead of JSON
- Refactoring dependency management
- Replacing pylint with Ruff
- added simpler app entrypoint
