+++ shared/constants.py
@@ -1,0 +1,18 @@
+# Magic 8-Ball Responses
+EIGHT_BALL_RESPONSES = [
+    "It is certain.", "It is decidedly so.", "Without a doubt.", "Yes - definitely.",
+    "You may rely on it.", "As I see it, yes.", "Most likely.", "Outlook good.",
+    "Yes.", "Signs point to yes.", "Reply hazy, try again.", "Ask again later.",
+    "Better not tell you now.", "Cannot predict now.", "Concentrate and ask again.",
+    "Don't count on it.", "My reply is no.", "My sources say no.", "Outlook not so good.",
+    "Very doubtful."
+]
+
+# Rock Paper Scissors Game Constants
+RPS_CHOICES = ['rock', 'paper', 'scissors']
+
+RPS_WINNING_CONDITIONS = {
+    'rock': 'scissors',
+    'paper': 'rock',
+    'scissors': 'paper'
+}
